64bit対応: int -> NSInteger に変更

+++ Classes/AppDelegate.h
@@ -25,7 +25,7 @@
 + (NSString *)appVersion;
 
 //+ (void)trackPageview:(NSString *)url;
-+ (void)trackEvent:(NSString *)category action:(NSString *)action label:(NSString *)label value:(int)value;
++ (void)trackEvent:(NSString *)category action:(NSString *)action label:(NSString *)label value:(NSInteger)value;
 
 // Utility
 #define _L(msg)  NSLocalizedString(msg, @"")
