+++ Classes/controllers/ConfigViewController.m
@@ -98,7 +98,7 @@
     if (config.cutoffDate == 0) {
         cutoffDateDescLabel.text = _L(@"End of month");
     } else {
-        cutoffDateDescLabel.text = [NSString stringWithFormat:@"%d", config.cutoffDate];
+        cutoffDateDescLabel.text = [NSString stringWithFormat:@"%ld", (long)config.cutoffDate];
     }
 
     currencyLabel.text = _L(@"Currency");
@@ -213,7 +213,7 @@
     }
 }
 
-- (BOOL)genSelectListViewChanged:(GenSelectListViewController *)vc identifier:(int)id
+- (BOOL)genSelectListViewChanged:(GenSelectListViewController *)vc identifier:(NSInteger)id
 {
     Config *config = [Config instance];
     NSString *currency = nil;
